refactor: decouple model resolution and prompt building logic into new modules
